requirements.txt
audioread (1.2.1)
awscli (1.8.12)
backports.ssl-match-hostname (3.4.0.2)
botocore (1.2.10)
certifi (14.5.14)
cffi (1.5.0)
characteristic (14.3.0)
colorama (0.3.3)
coverage (4.0.2)
cryptography (1.2.1)
decorator (3.4.2)
dj-database-url (0.3.0)
Django (1.9.3)
django-allauth (0.24.1)
django-braces (1.8.1)
django-extensions (1.6.1)
django-rest (0.0.1)
django-rest-auth (0.7.0)
django-static-precompiler (1.2)
djangorestframework (3.3.2)
docutils (0.12)
enum34 (1.1.2)
extraction (0.2)
functools32 (3.2.3.post2)
greenlet (0.4.9)
httplib2 (0.9)
hurry.filesize (0.9)
idna (2.0)
ipaddress (1.0.14)
Jinja2 (2.7.3)
jmespath (0.9.0)
joblib (0.8.4)
jsonschema (2.5.1)
lxml (3.5.0)
MarkupSafe (0.23)
mock (1.0.1)
nose (1.3.4)
oauthlib (1.0.3)
optional-django (0.1.0)
pefile (1.2.10.post114)
Pillow (3.1.0)
pip (8.0.0)
plumbum (1.6.1.post0)
protobuf (2.6.1)
psycopg2 (2.6.1)
pyasn1 (0.1.9)
pyasn1-modules (0.0.8)
pybonjour (1.1.1, /Users/patrickryan/Library/src/pybonjour)
pycparser (2.14)
pycrypto (2.6.1)
pygeoip (0.3.2)
PyInstaller (3.1.1)
pyOpenSSL (0.15.1)
pyparsing (2.0.3)
python-dateutil (2.2)
python-mimeparse (0.1.4)
python-openid (2.2.5)
python-ptrace (0.8.1)
pytz (2015.7)
PyYAML (3.11)
pyzmq (14.7.0)
queuelib (1.4.2)
requests (2.3.0)
requests-oauthlib (0.4.1)
rpyc (3.3.0)
rsa (3.2)
service-identity (14.0.0)
setuptools (12.0.5)
shortuuid (0.4.2)
six (1.10.0)
tornado (4.0)
Twisted (15.5.0)
tzlocal (1.2)
Unidecode (0.4.16)
virtualenv (13.1.2)
w3lib (1.13.0)
wheel (0.24.0)
whitenoise (2.0.6)
zope.interface (4.1.3)